requirements.txt
streamlit==1.54.0

langchain==1.2.10
langchain-core==1.2.13
langchain-community==0.4.1
langchain-openai==1.1.9
langchain-text-splitters==1.1.0

openai==2.17.0
tiktoken==0.12.0

faiss-cpu==1.13.2
sentence-transformers==5.2.2

msoffcrypto-tool==6.0.0
python-docx==1.2.0
openpyxl==3.1.5
msoffcrypto-tool==6.0.0
